+++ test/tests_ColVariable.cpp
@@ -19,6 +19,9 @@
 #include "ColVariable.h"
 #include "SMSTypedefs.h"
 
+// last, so that the headers above are read as the library was compiled
+#include "TestAssert.h"
+
 /*--------------------------------------------------------------------------*/
 /*-------------------------------- USING -----------------------------------*/
 /*--------------------------------------------------------------------------*/